planner: wait until binding is loaded into memory if binding cache miss
