+++ pkg/bindinfo/BUILD.bazel
@@ -41,6 +41,7 @@ go_library(
         "//pkg/util/table-filter",
         "@com_github_ngaut_pools//:pools",
         "@com_github_pingcap_errors//:errors",
+        "@com_github_pingcap_failpoint//:failpoint",
         "@org_golang_x_sync//singleflight",
         "@org_uber_go_zap//:zap",
     ],
