+++ pkg/bindinfo/tests/BUILD.bazel
@@ -2,14 +2,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_test")
 
 go_test(
     name = "tests_test",
-    timeout = "short",
+    timeout = "moderate",
     srcs = [
         "bind_test.go",
         "main_test.go",
     ],
     flaky = True,
     race = "on",
-    shard_count = 16,
+    shard_count = 18,
     deps = [
         "//pkg/bindinfo",
         "//pkg/bindinfo/internal",
@@ -23,6 +23,7 @@ go_test(
         "//pkg/util",
         "//pkg/util/parser",
         "//pkg/util/stmtsummary",
+        "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
         "@org_uber_go_goleak//:goleak",
     ],
