+++ pkg/planner/optimize.go
@@ -216,9 +216,13 @@ func Optimize(ctx context.Context, sctx sessionctx.Context, node ast.Node, is in
 	}
 
 	enableUseBinding := sessVars.UsePlanBaselines
-	bindHandle := domain.GetDomain(sctx).BindHandle()
 	stmtNode, isStmtNode := node.(ast.StmtNode)
-	bindings, match, scope := matchSQLBinding(sctx, stmtNode, bindHandle.LoadBindingsFromStorageToCache)
+	binding, match, scope := bindinfo.MatchSQLBinding(sctx, stmtNode)
+	var bindings bindinfo.Bindings
+	if match {
+		bindings = []bindinfo.Binding{binding}
+	}
+
 	useBinding := enableUseBinding && isStmtNode && match
 	if sessVars.StmtCtx.EnableOptimizerDebugTrace {
 		failpoint.Inject("SetBindingTimeToZero", func(val failpoint.Value) {
@@ -363,24 +367,6 @@ func Optimize(ctx context.Context, sctx sessionctx.Context, node ast.Node, is in
 	}
 
 	return bestPlan, names, nil
-}
-
-func tryMatchSQLBinding(sctx sessionctx.Context, stmtNode ast.StmtNode, syncFunc func(string)) (bindings bindinfo.Bindings, matched bool, scope string) {
-	bindings, match, scope := matchSQLBinding(sctx, stmtNode, syncFunc)
-	if match {
-		return
-	}
-	// try again.
-	bindings, match, scope = matchSQLBinding(sctx, stmtNode, syncFunc)
-	return
-}
-
-func matchSQLBinding(sctx sessionctx.Context, stmtNode ast.StmtNode, syncFunc func(string)) (bindings bindinfo.Bindings, matched bool, scope string) {
-	binding, match, scope := bindinfo.MatchSQLBinding(sctx, stmtNode, syncFunc)
-	if match {
-		bindings = []bindinfo.Binding{binding}
-	}
-	return
 }
 
 // OptimizeForForeignKeyCascade does optimization and creates a Plan for foreign key cascade.
